+++ source/ensure/index.ts
@@ -1,1 +1,1 @@
-export { Ensure } from "./Ensure.js";
+export { argumentIsProvided } from "./argumentIsProvided.js";
